+++ packages/trpc/src/internal/init.ts
@@ -93,10 +93,10 @@ export const authenticatedProcedureBuilder = baseProcedureBuilder.use(
 export const adminAuthenticatedProcedureBuilder =
   authenticatedProcedureBuilder.use(async ({ ctx, next }) => {
     if (ctx.session.user.role !== "ADMIN") {
-      throw new TRPCError({ code: "UNAUTHORIZED" });
+      throw new TRPCError({ code: "FORBIDDEN" });
     }
 
-    return next();
+    return next({ ctx });
   });
 
 // This middleware is used to prevent authenticated users from accessing a resource
